+++ Source/Krypton Ribbon Examples/Quick Access Toolbar/Form1.cs
@@ -55,19 +55,29 @@ namespace QuickAccessToolbar
         private void checkSetQATPosition_CheckedButtonChanged(object sender, EventArgs e)
         {
             if (checkSetQATPosition.CheckedButton == checkButtonAbove)
+            {
                 kryptonRibbon.QATLocation = QATLocation.Above;
+            }
             else if (checkSetQATPosition.CheckedButton == checkButtonBelow)
+            {
                 kryptonRibbon.QATLocation = QATLocation.Below;
+            }
             else if (checkSetQATPosition.CheckedButton == checkButtonHidden)
+            {
                 kryptonRibbon.QATLocation = QATLocation.Hidden;
+            }
         }
 
         private void checkSetQATUserChange_CheckedButtonChanged(object sender, EventArgs e)
         {
             if (checkSetQATUserChange.CheckedButton == checkButtonAllowUserChanges)
+            {
                 kryptonRibbon.QATUserChange = true;
+            }
             else if (checkSetQATUserChange.CheckedButton == checkButtonDisallowUserChanges)
+            {
                 kryptonRibbon.QATUserChange = false;
+            }
         }
 
         private void buttonAdd_Click(object sender, EventArgs e)
@@ -86,7 +96,9 @@ namespace QuickAccessToolbar
         private void buttonRemove_Click(object sender, EventArgs e)
         {
             if (kryptonRibbon.QATButtons.Count > 0)
+            {
                 kryptonRibbon.QATButtons.RemoveAt(0);
+            }
 
             UpdateButtons();
         }